refactor http api demo, extract common callback handler
